Added functional emojis button
Added a button with a popup when clicked. The popup menu has some emojis, when you click them, they will appear in the textbox as expected. At the moment, I couldn't make the emojis be colorful. It may require a full conversion of the WPF project to WinUI 3 / UWP / MAUI. Until then, emojis are monochrome.